worker/requirements.txt
google-cloud-pubsub
python-dotenv
openai   # solo si usarás OpenAI
google-cloud-firestore  # solo si guardarás en Firestore
